RDA View Page: fixed IE8/9 slowness (bug spreadsheet description follows):
ANDS internal suggested links load extremely slowly in IE8 and pretty slow in
IE9.
In IE8 it looks as if clicking on the link does nothing. IE9 is slightly
faster but still slow compared to FF and Chrome.
Maybe load popout with a loading icon/message

Resolution: bloody qtip plugin.... again! Employ an alternative strategy to
show the qtip (i.e. not with events) in a way that IE doesn't block rendering